+++ src/IntervalFormat.php
@@ -26,12 +26,10 @@ enum IntervalFormat
     case Iso8601DurationEnd;
     case Iso8601StartEnd;
     case Iso8601;
-    case Canonical;
 
     private const string REGEXP_ISO80000 = '/^\[(?<start>[^,)]*),(?<end>[^,)]*)\)$/';
     private const string REGEXP_BOURBAKI = '/^\[(?<start>[^,\[]*),(?<end>[^,\[]*)\[$/';
     private const string REGEXP_ISO8601 = '/^(?<start>[^\/]+)\/(?<end>[^\/]+)$/';
-    private const string REGEXP_CANONICAL = '/^(?<start>[^\/]+)\/(?<end>.+)\[\)$/';
 
     /**
      * @return non-empty-string
@@ -48,7 +46,6 @@ enum IntervalFormat
             self::Iso8601StartEnd => $start.'/'.$end,
             self::Iso80000 => '['.$start.','.$end.')',
             self::Bourbaki => '['.$start.','.$end.'[',
-            self::Canonical => $start.'/'.$interval->duration->format().'[)',
         };
     }
 
@@ -61,7 +58,6 @@ enum IntervalFormat
         $pattern = match ($this) {
             self::Bourbaki => self::REGEXP_BOURBAKI,
             self::Iso80000 => self::REGEXP_ISO80000,
-            self::Canonical => self::REGEXP_CANONICAL,
             default => self::REGEXP_ISO8601,
         };
 
@@ -176,7 +172,6 @@ enum IntervalFormat
     private function supportsStartDuration(): bool
     {
         return match ($this) {
-            self::Canonical,
             self::Iso8601,
             self::Iso8601StartDuration => true,
             default => false,